requirements.txt
google-cloud-storage==2.18.2
google-cloud-speech==2.27.0
google-cloud-videointelligence==2.13.4
google-cloud-vision==3.8.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.25.0
google-cloud-pubsub==2.23.0
pydantic==2.9.2
python-dotenv==1.0.1
ffmpeg-python==0.2.0
pytube==15.0.0
click==8.1.7
python-dateutil==2.9.0.post0
opencv-python-headless==4.10.0.84
pandas==2.2.3
numpy==1.26.4
google-generativeai==0.8.3
